# apps/api-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Lagrange Station API",
    "version": "1.0.0",
    "description": "Versioned REST/OpenAPI contract for Lagrange Station. Common rules (design §12.1): `/api/v1` prefix, JSON bodies, cursor pagination, `X-Request-Id` correlation in/out, the typed `{error:{code,message,request_id,details?}}` envelope, and Idempotency-Key on every mutating route. No database/NT/provider models leak into payloads."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
    "/api/v1/auth/session": {
      "get": {
        "operationId": "get__api_v1_auth_session",
        "summary": "GET /api/v1/auth/session",
        "tags": [
          "v1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "501": {
            "$ref": "#/components/responses/Error501"
          }
        },
        "x-lagrange": {
          "auth": {
            "required": true,
            "session": "opaque __Host-lagrange_session cookie"
          },
          "ownership": {
            "owner_only": false,
            "scope": "actor-scoped via RLS (foreign resources are indistinguishable from missing)"
          },
          "entitlement": {
            "use": null,
            "fail_closed": true
          },
          "idempotency": {
            "required": false,
            "note": "read-only"
          },
          "audit": {
            "writer": null
          },
          "cache": {
            "policy": "no-store",
            "reason": "authenticated per-user data is never shared"
          },
          "errors": [
            "SESSION_UNKNOWN",
            "SESSION_EXPIRED",
            "INTERNAL"
          ],
          "phase": "current"
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "post__api_v1_auth_logout",
        "summary": "POST /api/v1/auth/logout",
        "tags": [
          "v1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "501": {
            "$ref": "#/components/responses/Error501"
          }
        },
        "x-lagrange": {
          "auth": {
            "required": true,
            "session": "opaque __Host-lagrange_session cookie"
          },
          "ownership": {
            "owner_only": false,
            "scope": "actor-scoped via RLS (foreign resources are indistinguishable from missing)"
          },
          "entitlement": {
            "use": null,
            "fail_closed": true
          },
          "idempotency": {
            "required": false,
            "natural": true,
            "note": "idempotent by nature; no key required"
          },
          "audit": {
            "writer": "audit_writer (append-only)",
            "fields": "actor/time/target/before-after/reason/correlation_id"
          },
          "cache": {
            "policy": "no-store",
            "reason": "authenticated per-user data is never shared"
          },
          "errors": [
            "SESSION_UNKNOWN",
            "SESSION_EXPIRED",
            "INTERNAL",
            "CSRF_DENIED",
            "INVALID_PARAMETER",
            "PAYLOAD_TOO_LARGE"
          ],
          "phase": "current"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyBody"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/csrf": {
      "get": {
        "operationId": "get__api_v1_auth_csrf",
        "summary": "GET /api/v1/auth/csrf",
        "tags": [
          "v1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "501": {
            "$ref": "#/components/responses/Error501"
          }
        },
        "x-lagrange": {
          "auth": {
            "required": true,
            "session": "opaque __Host-lagrange_session cookie"
          },
          "ownership": {
            "owner_only": false,
            "scope": "actor-scoped via RLS (foreign resources are indistinguishable from missing)"
          },
          "entitlement": {
            "use": null,
            "fail_closed": true
          },
          "idempotency": {
            "required": false,
            "natural": true,
            "note": "idempotent by nature; no key required"
          },
          "audit": {
            "writer": null
          },
          "cache": {
            "policy": "no-store",
            "reason": "authenticated per-user data is never shared"
          },
          "errors": [
            "SESSION_UNKNOWN",
            "SESSION_EXPIRED",
            "INTERNAL",
            "CSRF_DENIED",
            "INVALID_PARAMETER",
            "PAYLOAD_TOO_LARGE"
          ],
          "phase": "current"
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyBody"
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/step-up-check": {
      "get": {
        "operationId": "get__api_v1_auth_step_up_check",
        "summary": "GET /api/v1/auth/step-up-check",
        "tags": [
          "v1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "501": {
            "$ref": "#/components/responses/Error501"
          }
        },
        "x-lagrange": {
          "auth": {
            "required": true,
            "session": "opaque __Host-lagrange_session cookie"
          },
          "ownership": {
            "owner_only": false,
            "scope": "actor-scoped via RLS (foreign resources are indistinguishable from missing)"
          },
          "entitlement": {
            "use": null,
            "fail_closed": true
          },
          "idempotency": {
            "required": false,
            "note": "read-only"
          },
          "audit": {
            "writer": "audit_writer (append-only)",
            "fields": "actor/time/target/before-after/reason/correlation_id"
          },
          "cache": {
            "policy": "no-store",
            "reason": "authenticated per-user data is never shared"
          },
          "errors": [
            "SESSION_UNKNOWN",
            "SESSION_EXPIRED",
            "INTERNAL"
          ],
          "phase": "current"
        }
      }
    },
    "/api/v1/strategies": {
      "get": {
        "operationId": "get__api_v1_strategies",
        "summary": "GET /api/v1/strategies",
        "tags": [
          "v1"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "501": {
            "$ref": "#/components/responses/Error501"
          }
        },
        "x-lagrange": {
          "auth": {
            "required": true,
            "session": "opaque __Host-lagrange_session cookie"
          },
          "ownership": {
            "owner_only": false,
            "scope": "actor-scoped via RLS (foreign resources are indistinguishable from missing)"
          },
          "entitlement": {
            "use": null,
            "fail_closed": true
          },
          "idempotency": {
            "required": false,
            "note": "read-only"
          },
          "audit": {
            "writer": null
          },
          "cache": {
            "policy": "no-store",
            "reason": "authenticated per-user data is never shared"
          },
          "errors": [
            "SESSION_UNKNOWN",
            "SESSION_EXPIRED",
            "INTERNAL"
          ],
          "phase": "current"
        }
      }
    },
    "/api/v1/strategies/{strategy_id}": {
      "get": {
        "operationId": "get__api_v1_strategies__strategy_id_",
        "summary": "GET /api/v1/strategies/{strategy_id}",
        "tags": [
          "v1"
        ],
        "parameters": [
          {
            "name": "strategy_id",
            "in": "path",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "413": {
            "$ref": "#/components/responses/Error413"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          },
          "501": {
            "$ref": "#/components/responses/Error501"
          }
        },
        "x-lagrange": {
          "auth": {
            "required": true,
            "session": "opaque __Host-lagrange_session cookie"
          },
          "ownership": {
            "owner_only": false,
            "scope": "actor-scoped via RLS (foreign resources are indistinguishable from missing)"
          },
          "entitlement": {
            "use": null,
            "fail_closed": true
          },
          "idempotency": {
            "required": false,
            "note": "read-only"
          },
          "audit": {
            "writer": null
          },
          "cache": {
            "policy": "no-store",
            "reason": "authenticated per-user data is never shared"
          },
          "errors": [
            "SESSION_UNKNOWN",
            "SESSION_EXPIRED",
            "INTERNAL"
          ],
          "phase": "current"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmptyBody": {
        "type": "object",
        "additionalProperties": false
      }
    },
    "responses": {
      "Error400": {
        "description": "400 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error401": {
        "description": "401 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error403": {
        "description": "403 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error404": {
        "description": "404 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error409": {
        "description": "409 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error413": {
        "description": "413 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error422": {
        "description": "422 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error429": {
        "description": "429 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error500": {
        "description": "500 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Error501": {
        "description": "501 typed error envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-lagrange_session",
        "description": "Opaque first-party session cookie issued by the auth router"
      }
    }
  },
  "security": [
    {
      "sessionCookie": []
    }
  ]
}